+++ src/components/selectors.ts
@@ -1,6 +1,4 @@
-import { E2ESelectors } from '@grafana/e2e-selectors';
-
-export const Components = {
+export const components = {
   ConfigEditor: {
     AccessToken: 'data-testid Access Token',
   },
@@ -25,12 +23,11 @@ export const Components = {
     Number: {
       input: 'Query editor number',
     },
+    Issues: {
+      timeFieldInput: 'Query editor issues time field',
+    },
   },
   AnnotationEditor: {
     container: 'Annotation editor container',
   },
-};
-
-export const selectors: { components: E2ESelectors<typeof Components> } = {
-  components: Components,
 };